Sort the original keys instead of the parsed floats
This fixes a bug where if your key is "2.0", we'd parse it as the float `2` and try to look up "2" in the dictionary which would fail. Instead do the parsing inside of `sort` which is a bit more expensive asymptotically but in practice still cheap or in many cases exactly as costly as before (two calls, since most people would specify just 2x and 3x images at this point).